Move Cluster-specific code to rpc.cluster, mostly
