+++ rpc/RPCDatabaseManager.java
@@ -28,78 +28,47 @@ import io.grpc.Channel;
 import io.grpc.StatusRuntimeException;
 
 import java.util.List;
-import java.util.Map;
 import java.util.function.Supplier;
 
 import static grakn.client.common.exception.ErrorMessage.Client.MISSING_DB_NAME;
 
-public class RPCDatabaseManager {
-    public static class Core implements Grakn.DatabaseManager {
-        private final GraknGrpc.GraknBlockingStub blockingGrpcStub;
+public class RPCDatabaseManager implements Grakn.DatabaseManager {
+    private final GraknGrpc.GraknBlockingStub blockingGrpcStub;
 
-        public Core(Channel channel) {
-            blockingGrpcStub = GraknGrpc.newBlockingStub(channel);
-        }
-
-        @Override
-        public boolean contains(String name) {
-            return request(() -> blockingGrpcStub.databaseContains(DatabaseProto.Database.Contains.Req.newBuilder().setName(nonNull(name)).build()).getContains());
-        }
-
-        @Override
-        public void create(String name) {
-            request(() -> blockingGrpcStub.databaseCreate(DatabaseProto.Database.Create.Req.newBuilder().setName(nonNull(name)).build()));
-        }
-
-        @Override
-        public void delete(String name) {
-            request(() -> blockingGrpcStub.databaseDelete(DatabaseProto.Database.Delete.Req.newBuilder().setName(nonNull(name)).build()));
-        }
-
-        @Override
-        public List<String> all() {
-            return request(() -> ImmutableList.copyOf(blockingGrpcStub.databaseAll(DatabaseProto.Database.All.Req.getDefaultInstance()).getNamesList()));
-        }
-
-        private String nonNull(String name) {
-            if (name == null) throw new GraknClientException(MISSING_DB_NAME);
-            return name;
-        }
-
-        private static <RES> RES request(Supplier<RES> req) {
-            try {
-                return req.get();
-            } catch (StatusRuntimeException e) {
-                throw GraknClientException.of(e);
-            }
-        }
+    public RPCDatabaseManager(Channel channel) {
+        blockingGrpcStub = GraknGrpc.newBlockingStub(channel);
     }
 
-    public static class Cluster implements Grakn.DatabaseManager {
-        private final Map<Address.Cluster.Server, RPCDatabaseManager.Core> databaseManagers;
+    @Override
+    public boolean contains(String name) {
+        return request(() -> blockingGrpcStub.databaseContains(DatabaseProto.Database.Contains.Req.newBuilder().setName(nonNull(name)).build()).getContains());
+    }
 
-        public Cluster(Map<Address.Cluster.Server, Core> databaseManagers) {
-            this.databaseManagers = databaseManagers;
-        }
+    @Override
+    public void create(String name) {
+        request(() -> blockingGrpcStub.databaseCreate(DatabaseProto.Database.Create.Req.newBuilder().setName(nonNull(name)).build()));
+    }
 
-        @Override
-        public boolean contains(String name) {
-            return databaseManagers.values().iterator().next().contains(name);
-        }
+    @Override
+    public void delete(String name) {
+        request(() -> blockingGrpcStub.databaseDelete(DatabaseProto.Database.Delete.Req.newBuilder().setName(nonNull(name)).build()));
+    }
 
-        @Override
-        public void create(String name) {
-            databaseManagers.values().forEach(dbMgr -> dbMgr.create(name));
-        }
+    @Override
+    public List<String> all() {
+        return request(() -> ImmutableList.copyOf(blockingGrpcStub.databaseAll(DatabaseProto.Database.All.Req.getDefaultInstance()).getNamesList()));
+    }
 
-        @Override
-        public void delete(String name) {
-            databaseManagers.values().forEach(dbMgr -> dbMgr.delete(name));
-        }
+    private String nonNull(String name) {
+        if (name == null) throw new GraknClientException(MISSING_DB_NAME);
+        return name;
+    }
 
-        @Override
-        public List<String> all() {
-            return databaseManagers.values().iterator().next().all();
+    private static <RES> RES request(Supplier<RES> req) {
+        try {
+            return req.get();
+        } catch (StatusRuntimeException e) {
+            throw GraknClientException.of(e);
         }
     }
 }
