+++ rpc/RPCSession.java
@@ -21,48 +21,31 @@ package grakn.client.rpc;
 
 import com.google.protobuf.ByteString;
 import grakn.client.Grakn;
-import grakn.client.GraknClient;
 import grakn.client.GraknOptions;
 import grakn.client.common.exception.GraknClientException;
 import grakn.protocol.GraknGrpc;
 import grakn.protocol.SessionProto;
-import grakn.protocol.cluster.DatabaseProto;
 import io.grpc.Channel;
 import io.grpc.StatusRuntimeException;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
 
-import java.util.Collection;
-import java.util.HashMap;
-import java.util.Map;
-import java.util.Objects;
 import java.util.Timer;
 import java.util.TimerTask;
-import java.util.concurrent.ConcurrentHashMap;
-import java.util.concurrent.ConcurrentMap;
 import java.util.concurrent.atomic.AtomicBoolean;
-import java.util.stream.Collectors;
 
 import static grakn.client.GraknProtoBuilder.options;
-import static grakn.client.common.exception.ErrorMessage.Client.CLUSTER_NO_PRIMARY_REPLICA_YET;
-import static grakn.client.common.exception.ErrorMessage.Client.CLUSTER_REPLICA_NOT_PRIMARY;
-import static grakn.client.common.exception.ErrorMessage.Client.CLUSTER_UNABLE_TO_CONNECT;
-import static grakn.client.common.exception.ErrorMessage.Client.UNABLE_TO_CONNECT;
-import static grakn.client.common.exception.ErrorMessage.Internal.UNEXPECTED_INTERRUPTION;
 
-public class RPCSession {
-    public static class Core implements Grakn.Session {
+public class RPCSession implements Grakn.Session {
+    private final Channel channel;
+    private final String database;
+    private final Type type;
+    private final ByteString sessionId;
+    private final AtomicBoolean isOpen;
+    private final Timer pulse;
+    private final GraknGrpc.GraknBlockingStub blockingGrpcStub;
 
-        private final Channel channel;
-        private final String database;
-        private final Type type;
-        private final ByteString sessionId;
-        private final AtomicBoolean isOpen;
-        private final Timer pulse;
-        private final GraknGrpc.GraknBlockingStub blockingGrpcStub;
-
-        public Core(GraknClient.Core client, String database, Type type, GraknOptions options) {
-            try {this.channel = client.channel();
+    public RPCSession(RPCClient client, String database, Type type, GraknOptions options) {
+        try {
+            this.channel = client.channel();
             this.database = database;
             this.type = type;
             blockingGrpcStub = GraknGrpc.newBlockingStub(channel);
@@ -78,348 +61,75 @@ public class RPCSession {
         }
     }
 
-        @Override
-        public Grakn.Transaction transaction(Grakn.Transaction.Type type) {
-            return transaction(type, GraknOptions.core());
-        }
+    @Override
+    public Grakn.Transaction transaction(Grakn.Transaction.Type type) {
+        return transaction(type, GraknOptions.core());
+    }
 
-        @Override
-        public Grakn.Transaction transaction(Grakn.Transaction.Type type, GraknOptions options) {
-            return new RPCTransaction(this, sessionId, type, options);
-        }
+    @Override
+    public Grakn.Transaction transaction(Grakn.Transaction.Type type, GraknOptions options) {
+        return new RPCTransaction(this, sessionId, type, options);
+    }
 
-        @Override
-        public Type type() {
-            return type;
-        }
+    @Override
+    public Type type() {
+        return type;
+    }
 
-        @Override
-        public boolean isOpen() {
-            return isOpen.get();
-        }
+    @Override
+    public boolean isOpen() {
+        return isOpen.get();
+    }
 
-        @Override
-        public void close() {
-            if (isOpen.compareAndSet(true, false)) {
-                pulse.cancel();
-                try {
-                    blockingGrpcStub.sessionClose(SessionProto.Session.Close.Req.newBuilder().setSessionId(sessionId).build());
-                } catch (StatusRuntimeException e) {
-                    throw GraknClientException.of(e);
-                }
-            }
-        }
-
-        @Override
-        public String database() {
-            return database;
-        }
-
-        Channel channel() { return channel; }
-
-        public void pulse() {
-            boolean alive;
+    @Override
+    public void close() {
+        if (isOpen.compareAndSet(true, false)) {
+            pulse.cancel();
             try {
-                alive = blockingGrpcStub.sessionPulse(
-                        SessionProto.Session.Pulse.Req.newBuilder().setSessionId(sessionId).build()).getAlive();
-            } catch (StatusRuntimeException exception) {
-                alive = false;
-            }
-            if (!alive) {
-                isOpen.set(false);
-                pulse.cancel();
-            }
-        }
-
-        private static SessionProto.Session.Type sessionType(Type type) {
-            switch (type) {
-                case DATA:
-                    return SessionProto.Session.Type.DATA;
-                case SCHEMA:
-                    return SessionProto.Session.Type.SCHEMA;
-                default:
-                    return SessionProto.Session.Type.UNRECOGNIZED;
-            }
-        }
-
-        private class PulseTask extends TimerTask {
-            @Override
-            public void run() {
-                if (!isOpen()) return;
-                pulse();
+                blockingGrpcStub.sessionClose(SessionProto.Session.Close.Req.newBuilder().setSessionId(sessionId).build());
+            } catch (StatusRuntimeException e) {
+                throw GraknClientException.of(e);
             }
         }
     }
 
-    public static class Cluster implements Grakn.Session {
-        private static final Logger LOG = LoggerFactory.getLogger(Cluster.class);
-        public static final int MAX_RETRY_PER_REPLICA = 10;
-        public static final int WAIT_FOR_PRIMARY_REPLICA_SELECTION_MS = 2000;
-        private final GraknClient.Cluster clusterClient;
-        private Database database;
-        private final String dbName;
-        private final Type type;
-        private final GraknOptions.Cluster options;
-        private final ConcurrentMap<Replica.Id, Core> coreSessions;
-        private boolean isOpen;
+    @Override
+    public String database() {
+        return database;
+    }
 
-        public Cluster(GraknClient.Cluster clusterClient, String database, Grakn.Session.Type type, GraknOptions.Cluster options) {
-            this.clusterClient = clusterClient;
-            this.dbName = database;
-            this.type = type;
-            this.options = options;
-            this.database = databaseDiscover();
-            coreSessions = new ConcurrentHashMap<>();
-            isOpen = true;
+    Channel channel() { return channel; }
+
+    public void pulse() {
+        boolean alive;
+        try {
+            alive = blockingGrpcStub.sessionPulse(
+                    SessionProto.Session.Pulse.Req.newBuilder().setSessionId(sessionId).build()).getAlive();
+        } catch (StatusRuntimeException exception) {
+            alive = false;
         }
+        if (!alive) {
+            isOpen.set(false);
+            pulse.cancel();
+        }
+    }
 
+    private static SessionProto.Session.Type sessionType(Type type) {
+        switch (type) {
+            case DATA:
+                return SessionProto.Session.Type.DATA;
+            case SCHEMA:
+                return SessionProto.Session.Type.SCHEMA;
+            default:
+                return SessionProto.Session.Type.UNRECOGNIZED;
+        }
+    }
+
+    private class PulseTask extends TimerTask {
         @Override
-        public Grakn.Transaction transaction(Grakn.Transaction.Type type) {
-            return transaction(type, GraknOptions.cluster());
-        }
-
-        @Override
-        public Grakn.Transaction transaction(Grakn.Transaction.Type type, GraknOptions options) {
-            GraknOptions.Cluster clusterOpt = options.asCluster();
-            if (clusterOpt.allowSecondaryReplica().isPresent() && clusterOpt.allowSecondaryReplica().get()) {
-                return transactionSecondaryReplica(type, clusterOpt);
-            } else {
-                return transactionPrimaryReplica(type, options);
-            }
-        }
-
-        @Override
-        public Type type() {
-            return type;
-        }
-
-        @Override
-        public boolean isOpen() {
-            return isOpen;
-        }
-
-        @Override
-        public void close() {
-            coreSessions.values().forEach(RPCSession.Core::close);
-            isOpen = false;
-        }
-
-        @Override
-        public String database() {
-            return dbName;
-        }
-
-        private Grakn.Transaction transactionPrimaryReplica(Grakn.Transaction.Type type, GraknOptions options) {
-            for (Replica replica: database.replicas()) {
-                int retry = 0;
-                while (retry < MAX_RETRY_PER_REPLICA) {
-                    try {
-                        Core primaryReplicaSession = coreSessions.computeIfAbsent(
-                                database.primaryReplica().id(),
-                                key -> {
-                                    LOG.debug("Opening a session to primary replica '{}'", key);
-                                    GraknClient.Core primaryReplicaClient = clusterClient.coreClient(key.address());
-                                    return primaryReplicaClient.session(key.database(), this.type, this.options);
-                                }
-                        );
-                        LOG.debug("Opening a transaction to primary replica '{}'", database.primaryReplica().id());
-                        return primaryReplicaSession.transaction(type, options);
-                    } catch (GraknClientException e) {
-                        retry++;
-                        if (e.getErrorMessage().equals(CLUSTER_REPLICA_NOT_PRIMARY)) {
-                            LOG.debug("Unable to open a session or transaction", e);
-                            database = databaseDiscover(replica.id().address());
-                        } else if (e.getErrorMessage().equals(CLUSTER_NO_PRIMARY_REPLICA_YET)) {
-                            LOG.debug("Unable to open a session or transaction", e);
-                            waitForPrimaryReplicaSelection();
-                            database = databaseDiscover(replica.id().address());
-                        } else if (e.getErrorMessage().equals(UNABLE_TO_CONNECT)) {
-                            LOG.debug("Unable to open a session or transaction", e);
-                            break;
-                        } else {
-                            throw e;
-                        }
-                    }
-                }
-            }
-            throw clusterNotAvailableException();
-        }
-
-        private Grakn.Transaction transactionSecondaryReplica(Grakn.Transaction.Type type, GraknOptions.Cluster options) {
-            for (Replica replica: database.replicas()) {
-                try {
-                    Core selectedSession = coreSessions.computeIfAbsent(
-                            replica.id(),
-                            key -> {
-                                LOG.debug("Opening a session to '{}'", key);
-                                return clusterClient.coreClient(key.address()).session(key.database(), this.type, this.options);
-                            }
-                    );
-                    LOG.debug("Opening read secondary transaction to secondary replica '{}'", replica);
-                    return selectedSession.transaction(type, options);
-                } catch (GraknClientException e) {
-                    if (e.getErrorMessage().equals(UNABLE_TO_CONNECT)) {
-                        LOG.debug("Unable to open a session or transaction to " + replica.id() + ". Reattempting to the next one.", e);
-                    } else {
-                        throw e;
-                    }
-                }
-            }
-            throw clusterNotAvailableException();
-        }
-
-        private Database databaseDiscover() {
-            for (Address.Cluster.Server server: clusterClient.clusterMembers()) {
-                try {
-                    return databaseDiscover(server);
-                } catch (StatusRuntimeException e) {
-                    LOG.debug("Unable to perform database discovery to " + server + ". Reattempting to the next one.", e);
-                }
-            }
-            throw clusterNotAvailableException();
-        }
-
-        private Database databaseDiscover(Address.Cluster.Server server) {
-            return Database.ofProto(
-                    clusterClient.graknClusterRPC(server).databaseDiscover(
-                        DatabaseProto.Database.Discover.Req.newBuilder()
-                                .setDatabase(dbName)
-                                .build()
-                    )
-            );
-        }
-
-        private GraknClientException clusterNotAvailableException() {
-            String addresses = clusterClient.clusterMembers().stream().map(Address.Cluster.Server::toString).collect(Collectors.joining(","));
-            return new GraknClientException(CLUSTER_UNABLE_TO_CONNECT, addresses); // remove ambiguity by casting to Object
-        }
-
-        private void waitForPrimaryReplicaSelection() {
-            try {
-                Thread.sleep(WAIT_FOR_PRIMARY_REPLICA_SELECTION_MS);
-            } catch (InterruptedException e2) {
-                throw new GraknClientException(UNEXPECTED_INTERRUPTION);
-            }
-        }
-
-        public static class Database {
-            private final Map<Replica.Id, Replica> replicas;
-
-            private Database(Map<Replica.Id, Replica> replicas) {
-                this.replicas = replicas;
-            }
-
-            public static Database ofProto(DatabaseProto.Database.Discover.Res res) {
-                Map<Replica.Id, Replica> replicaMap = new HashMap<>();
-
-                for (DatabaseProto.Database.Discover.Res.Replica replica: res.getReplicasList()) {
-                    Replica.Id id = new Replica.Id(Address.Cluster.Server.parse(replica.getAddress()), replica.getDatabase());
-                    replicaMap.put(id, Replica.ofProto(replica));
-                }
-
-                return new Database(replicaMap);
-            }
-
-            private Replica primaryReplica() {
-                Map.Entry<Replica.Id, Replica> initial = replicas.entrySet().iterator().next();
-                Map.Entry<Replica.Id, Replica> reduce = replicas.entrySet().stream()
-                        .filter(entry -> entry.getValue().isPrimary())
-                        .reduce(initial, (acc, e) -> e.getValue().term > acc.getValue().term ? e : acc);
-                if (reduce.getValue().isPrimary()) return reduce.getValue();
-                else throw new GraknClientException(CLUSTER_NO_PRIMARY_REPLICA_YET, (reduce.getValue().term()));
-            }
-
-            public Collection<Replica> replicas() {
-                return replicas.values();
-            }
-        }
-
-        public static class Replica {
-            private final Replica.Id id;
-            private final boolean isPrimary;
-            private final long term;
-
-            private Replica(Replica.Id id, long term, boolean isPrimary) {
-                this.id = id;
-                this.term = term;
-                this.isPrimary = isPrimary;
-            }
-
-            public static Replica ofProto(DatabaseProto.Database.Discover.Res.Replica replica) {
-                return new Replica(
-                        new Id(Address.Cluster.Server.parse(replica.getAddress()), replica.getDatabase()),
-                        replica.getTerm(),
-                        replica.getIsPrimary()
-                );
-            }
-
-            public Id id() {
-                return id;
-            }
-
-            public long term() {
-                return term;
-            }
-
-            public boolean isPrimary() {
-                return isPrimary;
-            }
-
-            @Override
-            public boolean equals(Object o) {
-                if (this == o) return true;
-                if (o == null || getClass() != o.getClass()) return false;
-                Replica replica = (Replica) o;
-                return term == replica.term &&
-                        isPrimary == replica.isPrimary;
-            }
-
-            @Override
-            public int hashCode() {
-                return Objects.hash(isPrimary, term);
-            }
-
-            @Override
-            public String toString() {
-                return id + ":" + (isPrimary ? "P" : "S") + ":" + term;
-            }
-
-            public static class Id {
-                private final Address.Cluster.Server address;
-                private final String database;
-
-                Id(Address.Cluster.Server address, String database) {
-                    this.address = address;
-                    this.database = database;
-                }
-
-                public Address.Cluster.Server address() {
-                    return address;
-                }
-
-                public String database() {
-                    return database;
-                }
-
-                @Override
-                public boolean equals(Object o) {
-                    if (this == o) return true;
-                    if (o == null || getClass() != o.getClass()) return false;
-                    Id id = (Id) o;
-                    return Objects.equals(address, id.address) &&
-                            Objects.equals(database, id.database);
-                }
-
-                @Override
-                public int hashCode() {
-                    return Objects.hash(address, database);
-                }
-
-                @Override
-                public String toString() {
-                    return address + "/" + database;
-                }
-            }
+        public void run() {
+            if (!isOpen()) return;
+            pulse();
         }
     }
 }
